refactor: 도메인(entity, dto, constant) 패키지 구조 리펙토링
