+++ src/main/java/com/grantip/backend/domain/calender/repository/CalenderRepository.java
@@ -1,6 +1,6 @@
 package com.grantip.backend.domain.calender.repository;
 
-import com.grantip.backend.domain.calender.entity.ScholarCalender;
+import com.grantip.backend.domain.calender.domain.entity.ScholarCalender;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
