+++ src/main/java/com/grantip/backend/domain/email/domain/dto/reqest/EmailRequest.java
@@ -1,4 +1,5 @@
-package com.grantip.backend.domain.email.dto;
+package com.grantip.backend.domain.email.domain.dto.reqest;
+
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -9,12 +10,9 @@ import lombok.AllArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class EmailVerify {
+public class EmailRequest {
     @Email(message = "올바른 이메일 형식이어야 합니다.")
     @NotBlank(message = "이메일은 필수 입력값입니다.")
     private String email;
-
-    @NotBlank(message = "인증 코드는 필수 입력값입니다.")
-    private String code;
 }
