+++ src/main/java/com/grantip/backend/domain/email/domain/dto/reqest/EmailVerify.java
@@ -1,5 +1,4 @@
-package com.grantip.backend.domain.email.dto;
-
+package com.grantip.backend.domain.email.domain.dto.reqest;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -10,9 +9,12 @@ import lombok.AllArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class EmailRequest {
+public class EmailVerify {
     @Email(message = "올바른 이메일 형식이어야 합니다.")
     @NotBlank(message = "이메일은 필수 입력값입니다.")
     private String email;
+
+    @NotBlank(message = "인증 코드는 필수 입력값입니다.")
+    private String code;
 }
