+++ src/main/java/com/grantip/backend/domain/token/domain/dto/TokenDto.java
@@ -1,4 +1,4 @@
-package com.grantip.backend.domain.token.dto;
+package com.grantip.backend.domain.token.domain.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -12,6 +12,5 @@ import lombok.Setter;
 public class TokenDto {
     private  String accessToken;
     private  String refreshToken;
-
 }
