+++ src/main/java/com/grantip/backend/domain/token/repository/TokenRepository.java
@@ -1,6 +1,6 @@
 package com.grantip.backend.domain.token.repository;
 
-import com.grantip.backend.domain.token.entity.RefreshToken;
+import com.grantip.backend.domain.token.domain.entity.RefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
