+++ src/main/java/com/grantip/backend/domain/user/domain/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.grantip.backend.domain.user.dto;
+package com.grantip.backend.domain.user.domain.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
