+++ src/main/java/com/grantip/backend/domain/user/domain/dto/request/VerifyPassword.java
@@ -1,4 +1,4 @@
-package com.grantip.backend.domain.user.dto;
+package com.grantip.backend.domain.user.domain.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
