+++ src/main/java/com/grantip/backend/domain/user/service/AuthService.java
@@ -3,18 +3,16 @@ package com.grantip.backend.domain.user.service;
 
 import com.grantip.backend.domain.email.repository.VerificationCodeRepository;
 import com.grantip.backend.domain.email.service.EmailService;
-import com.grantip.backend.domain.token.dto.TokenDto;
+import com.grantip.backend.domain.token.domain.dto.TokenDto;
 import com.grantip.backend.domain.token.service.TokenService;
-import com.grantip.backend.domain.user.dto.CustomUserDetails;
-import com.grantip.backend.domain.user.dto.LoginRequest;
-import com.grantip.backend.domain.user.dto.LoginResponse;
-import com.grantip.backend.domain.user.dto.SignupRequest;
-import com.grantip.backend.domain.user.entity.Role;
+import com.grantip.backend.domain.user.domain.dto.CustomUserDetails;
+import com.grantip.backend.domain.user.domain.dto.request.LoginRequest;
+import com.grantip.backend.domain.user.domain.dto.request.SignupRequest;
+import com.grantip.backend.domain.user.domain.constant.Role;
 import com.grantip.backend.global.code.ErrorCode;
 import com.grantip.backend.global.exception.CustomException;
-import com.grantip.backend.global.exception.EmailAuthException;
 import com.grantip.backend.global.util.JWTUtil;
-import com.grantip.backend.domain.user.entity.User;
+import com.grantip.backend.domain.user.domain.entity.User;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
